requirements.txt
certifi==2024.2.2
charset-normalizer==3.4.2
click==8.1.7
EbookLib==0.18
fuzzywuzzy==0.18.0
httpx==0.28.1
idna==3.10
jellyfish==1.2.0
Levenshtein==0.27.1
lxml==5.4.0
mutagen==1.47.0
PySimpleGUI==5.0.10
PyYAML==6.0.2
python-magic-bin==0.4.14
rapidfuzz==3.13.0
requests==2.32.4
tenacity==9.1.2
textdistance==4.6.3
